+++ docs/changelog.md
@@ -9,6 +9,7 @@
 - Add better linting when an action is modified inside an agent.
 - Add error message when an action from an agent inside Sema4ai folder is modified.
 - Add `kill-lock-holder` flag to action server to kill the lock holder when starting action server for doing OAuth2 flows.
+- Fixed issue with Publishing Action Package to Studio
 
 ## New in 2.12.0 (2025-05-07)